+++ packages/server/src/middlewares/LogRequestMiddleware.ts
@@ -1,24 +1,24 @@
 import { useContext } from "../context";
+import { createConsoleColors, createLogger, Logger, LogLevel } from "../logger";
 import { HttpCServerMiddleware } from "../server";
 
 
 type LogRequestMiddlewareOptions = {
     ansi?: boolean
+    level?: LogLevel
+    logger?: Logger
 }
 
 export function LogRequestMiddleware(options?: LogRequestMiddlewareOptions): HttpCServerMiddleware {
 
     const {
         ansi = true,
+        logger = createLogger(options),
     } = options || {}
 
-
-    const escape = (start: number, end: number, text: string) => `\x1b[${start}m${text}\x1b[${end}m`;
-    const identity = (x: string) => x;
-
-    const gray = ansi ? escape.bind(null, 90, 39) : identity;
-    const red = ansi ? escape.bind(null, 31, 39) : identity;
-    const green = ansi ? escape.bind(null, 32, 49) : identity;
+    const {
+        gray,
+    } = createConsoleColors(ansi);
 
 
     return async (call, next) => {
@@ -26,9 +26,9 @@ export function LogRequestMiddleware(options?: LogRequestMiddlewareOptions): Htt
 
         const elapsed = Date.now() - useContext().startedAt;
         if (result && result instanceof Error) {
-            console.log(`${red("ERROR")} ${gray(call.access)}\t${call.path} ${gray(`(${elapsed}ms)`)}`);
+            logger("error", `${gray(call.access)}\t${call.path} ${gray(`(${elapsed}ms)`)}`);
         } else {
-            console.log(`${green("SUCCESS")} ${gray(call.access)}\t${call.path} ${gray(`(${elapsed}ms)`)}`);
+            logger("success", `${gray(call.access)}\t${call.path} ${gray(`(${elapsed}ms)`)}`);
         }
 
         return result;